Move heap into subdirectory, and namespace.
